Adjust Hack mdelay time
I briefly discussed this with Mac - this is a hacky delay to allow the overclock to settle down before configuring the touchscreen. I think it might need to be longer to stop the GS in particular hanging on the Leapfrog screen on startup around 50% of the time.  May need some experimenting with values to see if this has any effect.